refactor | rename data dir to config
